% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dcast_getFormula}
\alias{dcast_getFormula}
\title{获取数据转换公式}
\usage{
dcast_getFormula(row_var, col_var)
}
\arguments{
\item{row_var}{行变量}

\item{col_var}{列变量}
}
\value{
返回公式
}
\description{
获取数据转换公式
}
\examples{
dcast_getFormula()
}
