% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctList.R
\name{gl_getAcctList}
\alias{gl_getAcctList}
\title{获取凭证明细}
\usage{
gl_getAcctList(conn = tsda::conn_rds("rpsrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取凭证明细
}
\examples{
gl_getAcctList()
}
