% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctList.R
\name{gl_getAcctSum}
\alias{gl_getAcctSum}
\title{获取凭证列表汇总数}
\usage{
gl_getAcctSum(conn = tsda::conn_rds("rpsrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取凭证列表汇总数
}
\examples{
gl_getAcctSum()
}
